split writep out, add unit test

+++ bk.h
@@ -5,7 +5,7 @@ struct BookKeeper;
 
 struct BookKeeper *BookKeeper_init();
 
-extern int N_blob;
+extern unsigned N_blob;
 
 
 /* allocate 'num' of pages. pad with 'content' 
@@ -26,7 +26,7 @@ int alloc_pages(struct BookKeeper *bk, unsigned num, char content);
 unsigned nr_allocated_pages(struct BookKeeper *bk);
 
 
-#define ROUND_UP(num, mod) ((((unsigned)num) + (mod) - 1) / (mod) * (mod))
+#define ROUND_UP(num, mod) ((((unsigned long long)num) + (mod) - 1) / (mod) * (mod))
 
 /*
  * return the real number of pages allocated
